+++ command/flow.go
@@ -1,12 +1,15 @@
-// +build darwin windows
-
 package command
 
 import (
+	"runtime"
+
 	"github.com/privatix/dapp-installer/pipeline"
 )
 
 func installFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return installLinuxFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedInstallFlags, nil),
 		newOperator("validate", validateToInstall, nil),
@@ -24,6 +27,9 @@ func installFlow() pipeline.Flow {
 }
 
 func updateFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return updateLinuxFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedUpdateFlags, nil),
 		newOperator("validate", checkInstallation, nil),
@@ -40,6 +46,9 @@ func updateFlow() pipeline.Flow {
 }
 
 func removeFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return removeLinuxFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedRemoveFlags, nil),
 		newOperator("validate", checkInstallation, nil),
@@ -60,6 +69,9 @@ func statusFlow() pipeline.Flow {
 }
 
 func installProductsFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return installLinuxProductsFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedInstallProductFlags, nil),
 		newOperator("validate", checkInstallation, nil),
@@ -68,6 +80,9 @@ func installProductsFlow() pipeline.Flow {
 }
 
 func updateProductsFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return updateLinuxProductsFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedUpdateProductFlags, nil),
 		newOperator("validate", checkInstallation, nil),
@@ -77,9 +92,85 @@ func updateProductsFlow() pipeline.Flow {
 }
 
 func removeProductsFlow() pipeline.Flow {
+	if runtime.GOOS == "linux" {
+		return removeLinuxProductsFlow()
+	}
 	return pipeline.Flow{
 		newOperator("processed flags", processedRemoveProductFlags, nil),
 		newOperator("validate", checkInstallation, nil),
 		newOperator("remove products", removeProducts, nil),
+	}
+}
+
+func installLinuxFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedInstallFlags, nil),
+		newOperator("validate", validateToInstall, nil),
+		newOperator("prepare", prepare, nil),
+		newOperator("init temp", initTemp, removeTemp),
+		newOperator("extract", extract, removeDapp),
+		newOperator("configure tor", installTor, nil),
+		newOperator("configure dapp", configureDapp, nil),
+		newOperator("install", installContainer, removeContainer),
+		newOperator("start", startContainer, stopContainer),
+		newOperator("create database", createDatabase, nil),
+		newOperator("install products", installProducts, removeProducts),
+		newOperator("finalize", finalize, nil),
+		newOperator("remove temp", removeTemp, nil),
+	}
+}
+
+func updateLinuxFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedUpdateFlags, nil),
+		newOperator("validate", checkContainer, nil),
+		newOperator("init temp", initTemp, removeTemp),
+		newOperator("stop", stopContainer, startContainer),
+		newOperator("update", updateContainer, restoreContainer),
+		//newOperator("disable daemons", disableDaemons, nil),
+		newOperator("start", startContainer, stopContainer),
+		newOperator("update database", updateDatabase, nil),
+		//newOperator("enable daemons", enableDaemons, nil),
+		newOperator("finalize", finalize, nil),
+		newOperator("remove temp", removeTemp, nil),
+		newOperator("remove backup", removeBackup, nil),
+	}
+}
+
+func removeLinuxFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedRemoveFlags, nil),
+		newOperator("validate", checkContainer, nil),
+		newOperator("stop", stopContainer, nil),
+		newOperator("remove", removeContainer, nil),
+		newOperator("remove dapp", removeDapp, nil),
+	}
+}
+
+func installLinuxProductsFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedInstallProductFlags, nil),
+		newOperator("validate", checkContainer, nil),
+		newOperator("install products", installProducts, removeProducts),
+		newOperator("finalize", finalize, nil),
+	}
+}
+
+func updateLinuxProductsFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedUpdateProductFlags, nil),
+		newOperator("validate", checkContainer, nil),
+		newOperator("update products", updateProducts, nil),
+		newOperator("start products", startProducts, nil),
+		newOperator("finalize", finalize, nil),
+	}
+}
+
+func removeLinuxProductsFlow() pipeline.Flow {
+	return pipeline.Flow{
+		newOperator("processed flags", processedRemoveProductFlags, nil),
+		newOperator("validate", checkContainer, nil),
+		newOperator("remove products", removeProducts, nil),
+		newOperator("finalize", finalize, nil),
 	}
 }
